% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_sample_rdirichlet.r
\name{generate_dirichlet_clusters_with_sampled_points}
\alias{generate_dirichlet_clusters_with_sampled_points}
\title{Draws from a Dirichlet distribution and shows the clusters
that were generated by this draw. Additionally, adds points to
these clusters and shows which clusters are occupied}
\usage{
generate_dirichlet_clusters_with_sampled_points(n, a, K)
}
\arguments{
\item{n}{Number of points to be drawn in the clusters}

\item{a}{Parameter that will be passed in to a Gamma distribution
in order to draw from the Dirichlet distribution.}

\item{K}{Number of clusters to draw}
}
\value{
No return value
}
\description{
Each point is generated one at a time, need to hit enter to generate
a new point. Typing "x" will stop the clustering and the function
will return.
}
\examples{
generate_dirichlet_clusters_with_sampled_points(15, 0.5, 20)
}
