% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\name{generate_split_data}
\alias{generate_split_data}
\title{Generates a dataset used to exemplify clustering
The cluster centers are set relatively far away to
see how well the algorithm performs in simple scenarios}
\usage{
generate_split_data(n, sd)
}
\arguments{
\item{n}{Number of datapoints to generate}

\item{sd}{Standard deviation from the cluster center}
}
\value{
Returns the datapoints and the cluster assignments.
The cluster assignments can be used to calculate the performance
of the clustering.
}
\description{
Generates a dataset used to exemplify clustering
The cluster centers are set relatively far away to
see how well the algorithm performs in simple scenarios
}
