% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dirichlet.r
\name{rdirichlet}
\alias{rdirichlet}
\title{Generate a sample from a Dirichlet distirbution}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{Number of observations.}

\item{alpha}{A vector containing the parameters for the Dirichlet distribution.}
}
\value{
A sample of n observations from the Dirichlet distribution.
}
\description{
Generate a sample from a Dirichlet distirbution
}
\examples{
rdirichlet(n=1, alpha=c(2, 2, 2))
}
