% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_dpmm.r
\name{rDPM_visual}
\alias{rDPM_visual}
\title{Sequentially generate draws from a Dirichlet process mixture model, by showing
step by step the iterations taken.
The plot is centered at 0, with x and y from -5 to 5.
The mixture draws the centres for clusters from a Normal distribution
with mean mu and standard deviation sigma_0}
\usage{
rDPM_visual(n, alpha, mu, sigma_0, sigma)
}
\arguments{
\item{n}{Number of observations.}

\item{alpha}{Alpha corresponding to GEM(alpha) used to draw the rho vector.}

\item{mu}{Mean of the Normal distribution used to draw the clusters.}

\item{sigma_0}{Standard deviation of the Normal distribution used
to draw the points around the cluster centre.}
}
\value{
Returns the n observations sampled from the DPMM distribution.
}
\description{
Hit enter to keep drawing until max n, type x to exit.
}
\examples{
rDPM_visual(n=30, alpha=3, mu=0, sigma_0=1.5, sigma=0.7)
}
