% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dirichlet.r
\name{rdirichlet}
\alias{rdirichlet}
\title{Add together two numbers}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{Number of observations.}

\item{alpha}{A vector containing the parameters for the Dirichlet distribution.}
}
\value{
A sample of n observations from the Dirichlet distribution.
}
\description{
Add together two numbers
}
\examples{
rdirichlet(n=1, alpha=c(2, 2, 2))
}
